Fix golint warning of funcutil
