+++ internal/util/funcutil/func.go
@@ -16,9 +16,14 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"io/ioutil"
 	"net"
+	"net/http"
 	"strconv"
 	"time"
+
+	"github.com/milvus-io/milvus/internal/log"
+	"go.uber.org/zap"
 
 	"github.com/go-basic/ipv4"
 	"github.com/milvus-io/milvus/internal/types"
@@ -28,6 +33,7 @@ import (
 	"github.com/milvus-io/milvus/internal/proto/internalpb"
 )
 
+// CheckGrpcReady wait for context timeout, or wait 100ms then send nil to targetCh
 func CheckGrpcReady(ctx context.Context, targetCh chan error) {
 	select {
 	case <-time.After(100 * time.Millisecond):
@@ -37,6 +43,7 @@ func CheckGrpcReady(ctx context.Context, targetCh chan error) {
 	}
 }
 
+// CheckPortAvailable check if a port is available to be listened on
 func CheckPortAvailable(port int) bool {
 	addr := ":" + strconv.Itoa(port)
 	listener, err := net.Listen("tcp", addr)
@@ -46,6 +53,7 @@ func CheckPortAvailable(port int) bool {
 	return err == nil
 }
 
+// GetAvailablePort return an available port that can be listened on
 func GetAvailablePort() int {
 	listener, err := net.Listen("tcp", ":0")
 	if err != nil {
@@ -56,10 +64,12 @@ func GetAvailablePort() int {
 	return listener.Addr().(*net.TCPAddr).Port
 }
 
+// GetLocalIP return the local ip address
 func GetLocalIP() string {
 	return ipv4.LocalIP()
 }
 
+// WaitForComponentStates wait for component's state to be one of the specific states
 func WaitForComponentStates(ctx context.Context, service types.Component, serviceName string, states []internalpb.StateCode, attempts uint, sleep time.Duration) error {
 	checkFunc := func() error {
 		resp, err := service.GetComponentStates(ctx)
@@ -87,18 +97,22 @@ func WaitForComponentStates(ctx context.Context, service types.Component, servic
 	return retry.Do(ctx, checkFunc, retry.Attempts(attempts), retry.Sleep(sleep))
 }
 
+// WaitForComponentInitOrHealthy wait for component's state to be initializing or healthy
 func WaitForComponentInitOrHealthy(ctx context.Context, service types.Component, serviceName string, attempts uint, sleep time.Duration) error {
 	return WaitForComponentStates(ctx, service, serviceName, []internalpb.StateCode{internalpb.StateCode_Initializing, internalpb.StateCode_Healthy}, attempts, sleep)
 }
 
+// WaitForComponentInitOrHealthy wait for component's state to be initializing
 func WaitForComponentInit(ctx context.Context, service types.Component, serviceName string, attempts uint, sleep time.Duration) error {
 	return WaitForComponentStates(ctx, service, serviceName, []internalpb.StateCode{internalpb.StateCode_Initializing}, attempts, sleep)
 }
 
+// WaitForComponentInitOrHealthy wait for component's state to be healthy
 func WaitForComponentHealthy(ctx context.Context, service types.Component, serviceName string, attempts uint, sleep time.Duration) error {
 	return WaitForComponentStates(ctx, service, serviceName, []internalpb.StateCode{internalpb.StateCode_Healthy}, attempts, sleep)
 }
 
+// ParseIndexParamsMap parse the jsonic index parameters to map
 func ParseIndexParamsMap(mStr string) (map[string]string, error) {
 	buffer := make(map[string]interface{})
 	err := json.Unmarshal([]byte(mStr), &buffer)
@@ -111,4 +125,62 @@ func ParseIndexParamsMap(mStr string) (map[string]string, error) {
 		ret[key] = valueStr
 	}
 	return ret, nil
+}
+
+const (
+	PulsarMaxMessageSizeKey = "maxMessageSize"
+)
+
+// GetPulsarConfig get pulsar configuration using pulsar admin api
+func GetPulsarConfig(protocol, ip, port, url string, args ...int64) (map[string]interface{}, error) {
+	var resp *http.Response
+	var err error
+
+	getResp := func() error {
+		log.Debug("function util", zap.String("url", protocol+"://"+ip+":"+port+url))
+		resp, err = http.Get(protocol + "://" + ip + ":" + port + url)
+		return err
+	}
+
+	var attempt uint = 10
+	var interval time.Duration = time.Second
+	if len(args) > 0 && args[0] > 0 {
+		attempt = uint(args[0])
+	}
+	if len(args) > 1 && args[1] > 0 {
+		interval = time.Duration(args[1])
+	}
+
+	err = retry.Do(context.TODO(), getResp, retry.Attempts(attempt), retry.Sleep(interval))
+	if err != nil {
+		log.Debug("failed to get config", zap.String("error", err.Error()))
+		return nil, err
+	}
+
+	defer resp.Body.Close()
+
+	body, err := ioutil.ReadAll(resp.Body)
+	log.Debug("get config", zap.String("config", string(body)))
+	if err != nil {
+		return nil, err
+	}
+
+	ret := make(map[string]interface{})
+	err = json.Unmarshal(body, &ret)
+	if err != nil {
+		return nil, err
+	}
+
+	return ret, nil
+}
+
+// GetAttrByKeyFromRepeatedKV return the value corresponding to key in kv pair
+func GetAttrByKeyFromRepeatedKV(key string, kvs []*commonpb.KeyValuePair) (string, error) {
+	for _, kv := range kvs {
+		if kv.Key == key {
+			return kv.Value, nil
+		}
+	}
+
+	return "", errors.New("key " + key + " not found")
 }
