+++ internal/util/funcutil/func_test.go
@@ -15,10 +15,13 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
+	"fmt"
+	"net/http"
 	"strconv"
 	"testing"
 	"time"
 
+	"github.com/jarcoal/httpmock"
 	"github.com/milvus-io/milvus/internal/proto/commonpb"
 	"github.com/milvus-io/milvus/internal/proto/internalpb"
 	"github.com/milvus-io/milvus/internal/proto/milvuspb"
@@ -196,4 +199,67 @@ func Test_ParseIndexParamsMap(t *testing.T) {
 	invalidStr := "invalid string"
 	_, err = ParseIndexParamsMap(invalidStr)
 	assert.NotEqual(t, err, nil)
+}
+
+func TestGetPulsarConfig(t *testing.T) {
+	httpmock.Activate()
+	defer httpmock.DeactivateAndReset()
+
+	runtimeConfig := make(map[string]interface{})
+	runtimeConfig[PulsarMaxMessageSizeKey] = strconv.FormatInt(5*1024*1024, 10)
+
+	protocol := "http"
+	ip := "pulsar"
+	port := "18080"
+	url := "/admin/v2/brokers/configuration/runtime"
+	httpmock.RegisterResponder("GET", protocol+"://"+ip+":"+port+url,
+		func(req *http.Request) (*http.Response, error) {
+			return httpmock.NewJsonResponse(200, runtimeConfig)
+		},
+	)
+
+	ret, err := GetPulsarConfig(protocol, ip, port, url)
+	assert.Equal(t, nil, err)
+	assert.Equal(t, len(ret), len(runtimeConfig))
+	assert.Equal(t, len(ret), 1)
+	for key, value := range ret {
+		assert.Equal(t, fmt.Sprintf("%v", value), fmt.Sprintf("%v", runtimeConfig[key]))
+	}
+}
+
+func TestGetPulsarConfig_Error(t *testing.T) {
+	protocol := "http"
+	ip := "pulsar"
+	port := "17777"
+	url := "/admin/v2/brokers/configuration/runtime"
+
+	ret, err := GetPulsarConfig(protocol, ip, port, url, 1, 1)
+	assert.NotNil(t, err)
+	assert.Nil(t, ret)
+}
+
+func TestGetAttrByKeyFromRepeatedKV(t *testing.T) {
+	kvs := []*commonpb.KeyValuePair{
+		{Key: "Key1", Value: "Value1"},
+		{Key: "Key2", Value: "Value2"},
+		{Key: "Key3", Value: "Value3"},
+	}
+
+	cases := []struct {
+		key      string
+		kvs      []*commonpb.KeyValuePair
+		value    string
+		errIsNil bool
+	}{
+		{"Key1", kvs, "Value1", true},
+		{"Key2", kvs, "Value2", true},
+		{"Key3", kvs, "Value3", true},
+		{"other", kvs, "", false},
+	}
+
+	for _, test := range cases {
+		value, err := GetAttrByKeyFromRepeatedKV(test.key, test.kvs)
+		assert.Equal(t, test.value, value)
+		assert.Equal(t, test.errIsNil, err == nil)
+	}
 }
